deps: DEP-TS - the width of `child_count` in tree-sitter 0.27, three legs for parser-specific invalidation, the archify gate keeps only one reader
- it/hs_grammar_pin.rs: `child_count()` is already u32 in 0.27, so
  `as u32` was removed
- it/trend_parser.rs (new): trend rows carrying the old stamp survive a
  parser revision, `run(…, Some(0))` reports pending 2 / rows 0, and
  after a full run pending is 0 and the row count equals the first
  measurement - seed_two_commits / run_all from `trend_rebuild.rs`
  promoted to pub(crate) and shared
- unit/dedup/parser.rs (new): a parser-revision mismatch -> the 13
  derived tables are cleared, trend rows and the meta other than
  full_build are kept, the epoch is rolled, `mark_full_build(old)` is
  refused, and reopening with the same key keeps the bag rows; a BEFORE
  DELETE trigger makes the whole invalidation roll back when it fails
- it/docs_diagrams.rs: the second reader of the archify cache,
  `cache_head`, was deleted (under a hollow cache `git -C
  cli/target/archify rev-parse HEAD` answers with CodeEraser's HEAD), so
  the named refusal on exit 2 of `node scripts/diagram.mjs --check` is
  the only predicate; the header comment changed with it, 169 -> 151
  lines

codex gpt-6-astra landed the code (the first three items), Claude
reviewed and landed the fourth; lib 373, it targeted legs 122 + eval 22
(the tokens.rs rows re-frozen by name under the EVAL-SET.md revival
protocol), clippy + fmt clean; six legs 984 / 119 / 0.

+++ it/hs_grammar_pin.rs
@@ -22,7 +22,7 @@ fn haskell_grammar_speaks_the_pinned_core_abi() {
     let mut stack = vec![root];
     while let Some(n) = stack.pop() {
         kinds.push(n.kind());
-        for i in (0..n.child_count() as u32).rev() {
+        for i in (0..n.child_count()).rev() {
             stack.push(n.child(i).unwrap());
         }
     }
